requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
pyyaml==6.0.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.35.0
apscheduler==3.10.4
sqlalchemy==2.0.30
great-expectations==0.18.15
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
plotly==5.22.0
altair==5.3.0
rich==13.7.1
tabulate==0.9.0
